Added a class called zonnescherm and made some changes in tuin and TuinDomotica
I created a class called Zonnescherm, it can be turn on or off, it checks the weather and if want u can check if it is raining or if there is any daylight.

I added a method called domoticaRun that runs all
device that are connect to the home automation by check the the mode.